+++ api/adobe_vendor_id.py
@@ -32,6 +32,7 @@ from core.model import (
     DelegatedPatronIdentifier,
     ExternalIntegration,
     Library,
+    Patron,
 )
 from core.scripts import Script
 
@@ -853,6 +854,32 @@ class AuthdataUtility(object):
         if not 'sub' in decoded:
             raise jwt.exceptions.DecodeError("No subject specified.")
         return library_uri, decoded['sub']
+
+    @classmethod
+    def _adobe_patron_identifier(self, patron):
+        _db = Session.object_session(patron)
+        internal = DataSource.lookup(_db, DataSource.INTERNAL_PROCESSING)
+
+        def refresh(credential):
+            credential.credential = str(uuid.uuid1())
+        patron_identifier = Credential.lookup(
+            _db, internal, AuthdataUtility.ADOBE_ACCOUNT_ID_PATRON_IDENTIFIER, patron,
+            refresher_method=refresh, allow_persistent_token=True
+        )
+        return patron_identifier.credential
+
+
+    def short_client_token_for_patron(self, patron_information):
+        if isinstance(patron_information, Patron):
+            # Find the patron's identifier for Adobe ID purposes.
+            patron_identifier = self._adobe_patron_identifier(
+                patron_information
+            )
+        else:
+            patron_identifier = patron_information
+
+        vendor_id, token = self.encode_short_client_token(patron_identifier)
+        return vendor_id, token
 
     def encode_short_client_token(self, patron_identifier):
         """Generate a short client token suitable for putting in an OPDS feed,
