+++ api/opds.py
@@ -923,18 +923,18 @@ class LibraryAnnotator(CirculationManagerAnnotator):
         link_tag.attrib.update(dict(href=fulfill_url))
         return link_tag
 
-    @classmethod
-    def _adobe_patron_identifier(self, patron):
-        _db = Session.object_session(patron)
-        internal = DataSource.lookup(_db, DataSource.INTERNAL_PROCESSING)
-
-        def refresh(credential):
-            credential.credential = str(uuid.uuid1())
-        patron_identifier = Credential.lookup(
-            _db, internal, AuthdataUtility.ADOBE_ACCOUNT_ID_PATRON_IDENTIFIER, patron,
-            refresher_method=refresh, allow_persistent_token=True
-        )
-        return patron_identifier.credential
+    # @classmethod
+    # def _adobe_patron_identifier(self, patron):
+    #     _db = Session.object_session(patron)
+    #     internal = DataSource.lookup(_db, DataSource.INTERNAL_PROCESSING)
+    #
+    #     def refresh(credential):
+    #         credential.credential = str(uuid.uuid1())
+    #     patron_identifier = Credential.lookup(
+    #         _db, internal, AuthdataUtility.ADOBE_ACCOUNT_ID_PATRON_IDENTIFIER, patron,
+    #         refresher_method=refresh, allow_persistent_token=True
+    #     )
+    #     return patron_identifier.credential
 
     def drm_device_registration_tags(self, license_pool, active_loan,
                                      delivery_mechanism):
@@ -976,27 +976,15 @@ class LibraryAnnotator(CirculationManagerAnnotator):
             cache_key = patron_identifier
         cached = self._adobe_id_tags.get(cache_key)
         if cached is None:
-            if isinstance(patron_identifier, Patron):
-                # Find the patron's identifier for Adobe ID purposes.
-                patron_identifier = self._adobe_patron_identifier(
-                    patron_identifier
-                )
             cached = []
             authdata = AuthdataUtility.from_config(self.library)
             if authdata:
-                # TODO: We would like to call encode() here, and have
-                # the client use a JWT as authdata, but we can't,
-                # because there's no way to use authdata to deactivate
-                # a device. So we've used this alternate technique
-                # that's much smaller than a JWT and can be smuggled
-                # into username/password.
-                vendor_id, jwt = authdata.encode_short_client_token(patron_identifier)
-
+                vendor_id, token = authdata.short_client_token_for_patron(patron_identifier)
                 drm_licensor = OPDSFeed.makeelement("{%s}licensor" % OPDSFeed.DRM_NS)
                 vendor_attr = "{%s}vendor" % OPDSFeed.DRM_NS
                 drm_licensor.attrib[vendor_attr] = vendor_id
                 patron_key = OPDSFeed.makeelement("{%s}clientToken" % OPDSFeed.DRM_NS)
-                patron_key.text = jwt
+                patron_key.text = token
                 drm_licensor.append(patron_key)
 
                 # Add the link to the DRM Device Management Protocol
